# For perf, remove Invoker pattern for fields