--- a/src/coreclr/vm/object.h
+++ b/src/coreclr/vm/object.h
@@ -1153,7 +1153,6 @@ protected:
     INT32               m_empty2;
     OBJECTREF           m_empty3;
     OBJECTREF           m_empty4;
-    OBJECTREF           m_empty5;
     FieldDesc *         m_pFD;
 
 public: