Fixes https://github.com/dotnet/runtime/issues/74550 by removing overhead added by the Invoker pattern which is unnecessary until IL emit support is added for fields.

Brings field getter to even with 6.0, and setter about 10% faster.

Suggest porting this to 7.0

For 8.0, this change will eventually be reverted and the corresponding IL emitted for field access.

Before:
```
|                                               Method |        Job |                                                                                                   Toolchain |       Mean |     Error |     StdDev |     Median |        Min |        Max | Ratio | RatioSD |  Gen 0 | Allocated | Alloc Ratio |
|----------------------------------------------------- |----------- |------------------------------------------------------------------------------------------------------------ |-----------:|----------:|-----------:|-----------:|-----------:|-----------:|------:|--------:|-------:|----------:|------------:|
|                                        Field_Get_int | Job-BQDNCA | \runtime60\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.9\corerun.exe |  30.993 ns | 0.1622 ns |  0.1438 ns |  30.972 ns |  30.768 ns |  31.259 ns |  0.87 |    0.01 | 0.0022 |      24 B |        1.00 |
|                                        Field_Get_int | Job-RQCXYM |   \runtime\artifacts\bin\testhost\net7.0-windows-Release-x64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe |  35.749 ns | 0.2806 ns |  0.2487 ns |  35.752 ns |  35.461 ns |  36.344 ns |  1.00 |    0.00 | 0.0021 |      24 B |        1.00 |
|                                                      |            |                                                                                                             |            |           |            |            |            |            |       |         |        |           |             |
|                                        Field_Set_int | Job-BQDNCA | \runtime60\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.9\corerun.exe |  36.037 ns | 0.2146 ns |  0.1903 ns |  36.018 ns |  35.744 ns |  36.418 ns |  0.98 |    0.01 | 0.0023 |      24 B |        1.00 |
|                                        Field_Set_int | Job-RQCXYM |   \runtime\artifacts\bin\testhost\net7.0-windows-Release-x64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe |  36.766 ns | 0.2736 ns |  0.2425 ns |  36.666 ns |  36.378 ns |  37.268 ns |  1.00 |    0.00 | 0.0023 |      24 B |        1.00 |
```

After
```
|                                               Method |        Job |                                                                                                   Toolchain |       Mean |     Error |    StdDev |     Median |        Min |        Max | Ratio | RatioSD |  Gen 0 | Allocated | Alloc Ratio |
|----------------------------------------------------- |----------- |------------------------------------------------------------------------------------------------------------ |-----------:|----------:|----------:|-----------:|-----------:|-----------:|------:|--------:|-------:|----------:|------------:|
|                                        Field_Get_int | Job-ELBTUN | \runtime60\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.9\corerun.exe |  34.137 ns | 0.5458 ns | 0.4839 ns |  34.074 ns |  33.391 ns |  35.035 ns |  1.00 |    0.01 | 0.0022 |      24 B |        1.00 |
|                                        Field_Get_int | Job-XYBTNI |   \runtime\artifacts\bin\testhost\net7.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |  34.164 ns | 0.0963 ns | 0.0854 ns |  34.169 ns |  33.972 ns |  34.275 ns |  1.00 |    0.00 | 0.0022 |      24 B |        1.00 |
|                                                      |            |                                                                                                             |            |           |           |            |            |            |       |         |        |           |             |
|                                        Field_Set_int | Job-ELBTUN | \runtime60\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.9\corerun.exe |  37.470 ns | 0.9993 ns | 1.1107 ns |  36.990 ns |  36.026 ns |  40.156 ns |  1.13 |    0.04 | 0.0022 |      24 B |        1.00 |
|                                        Field_Set_int | Job-XYBTNI |   \runtime\artifacts\bin\testhost\net7.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\corerun.exe |  33.512 ns | 0.1265 ns | 0.1056 ns |  33.506 ns |  33.369 ns |  33.780 ns |  1.00 |    0.00 | 0.0022 |      24 B |        1.00 |
```