@steveharter 2022-08-29
> @steveharter Do we have appropriate coverage in the perf repo (I saw this was customer reported)

Yes the existing (but new) benchmarks in the perf repo cover this. One item to consider however is that they are currently only based on an `int` field and thus could be expanded to reference types.